♻️ Refactor: info 페이지 서버페이지로 구현, MyInfoClient로 클라이언트 컴포넌트로 분리
