+++ src/features/user/ui/MyInfoPageHeader.tsx
@@ -1,0 +1,22 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import PageHeader from '@/shared/ui/page-header/PageHeader';
+
+/**
+ * 내 정보 페이지 헤더 컴포넌트
+ *
+ * @description
+ * `useRouter`를 사용하므로 클라이언트 컴포넌트로 분리합니다.
+ */
+export default function MyInfoPageHeader() {
+  const router = useRouter();
+
+  return (
+    <PageHeader
+      title="내 정보"
+      description="닉네임과 비밀번호를 수정하실 수 있습니다."
+      onBack={() => router.back()}
+    />
+  );
+}
